+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 * 【extra  】     增加ValidationUtil（pr#207@Gitee）
 * 【core   】     反射调用支持传递参数的值为null（pr#1205@Github）
 * 【core   】     HexUtil增加format方法（issue#I245NF@Gitee）
+* 【poi    】     ExcelWriter增加setCurrentRowToEnd方法（issue#I24A2R@Gitee）
+* 【core   】     ExcelWriter增加setCurrentRowToEnd方法（issue#I24A2R@Gitee）
 
 ### Bug修复
 * 【core   】     修复DateUtil.current使用System.nanoTime的问题（issue#1198@Github）